refactor: extract savings rate chart
